# Improve checkpoint series iterator.