… pkg/ingester/checkpoint_test.go
… pkg/ingester/encoding_test.go
… pkg/ingester/flush_test.go