This is a big refactoring allowing to use less memory when iterating over series for checkpointing purpose.

```
❯ benchcmp  before.txt after.txt8
benchmark                       old ns/op     new ns/op     delta
Benchmark_SeriesIterator-16     13062511      7110916       -45.56%

benchmark                       old allocs     new allocs     delta
Benchmark_SeriesIterator-16     28517          240            -99.16%

Improves memory usage of checkpointer series iterator.
benchmark                       old bytes     new bytes     delta
Benchmark_SeriesIterator-16     33072200      145433        -99.56%
```